DL-6 Add event assets, adjust wraping on event list, separate button from event to layout components
