+++ src/features/EventSection/components/EventList.tsx
@@ -1,28 +1,28 @@
 import styled from 'styled-components'
 import SingleEvent from './SingleEvent'
+import poster1 from '../../../assets/AukcjaTurtleUnite.jpg'
+import poster2 from '../../../assets/PierwszaAukjaTokenow.jpg'
+import poster3 from '../../../assets/plakat2AukcjaTokenow.jpg'
+import poster4 from '../../../assets/SummerArtBreak.jpg'
+import poster5 from '../../../assets/SummerArtBreak2.jpg'
 
 const EventList = () => {
   return (
     <EventListContainer>
-      <SingleEvent
-        alt='poster-1'
-        imgUrl='https://cdn.discordapp.com/attachments/608049459586269206/1037062861429669908/Summer_art_break.jpg'
-      />
-      <SingleEvent
-        alt='poster-2'
-        imgUrl='https://cdn.discordapp.com/attachments/608049459586269206/1037062861429669908/Summer_art_break.jpg'
-      />
-      <SingleEvent
-        alt='poster-3'
-        imgUrl='https://cdn.discordapp.com/attachments/608049459586269206/1037062861429669908/Summer_art_break.jpg'
-      />
+      <SingleEvent alt='poster-1' imgSrc={poster1} />
+      <SingleEvent alt='poster-2' imgSrc={poster2} />
+      <SingleEvent alt='poster-3' imgSrc={poster3} />
+      <SingleEvent alt='poster-4' imgSrc={poster4} />
+      <SingleEvent alt='poster-5' imgSrc={poster5} />
     </EventListContainer>
   )
 }
 
 const EventListContainer = styled.div`
   display: flex;
-  gap: 2rem;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 4rem 2rem;
 `
 
 export default EventList
