+++ src/features/EventSection/index.tsx
@@ -20,6 +20,7 @@ const Section = styled.section`
   gap: 1rem;
   align-items: center;
   justify-content: center;
+  padding-top: 8rem;
   background-color: #000000;
 `
