+++ src/layouts/Button/Button.tsx
@@ -1,16 +1,16 @@
 import styled from 'styled-components'
 
 export const Button = styled.button`
-  color: #ffffff;
-  background-color: #ffa500;
+  color: #0a0019;
+  background-color: #ffffff;
   border-radius: 6px;
   padding: 0.75rem 1.75rem;
   cursor: pointer;
 
   &:hover {
-    background-color: #e59400;
+    background-color: #ccc;
   }
   &:active {
-    background-color: #ffa500;
+    background-color: #0a0019;
   }
 `
