+++ src/layouts/Drawer/index.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react'
 import Box from '@mui/material/Box'
+import styled from 'styled-components'
 import SwipeableDrawer from '@mui/material/SwipeableDrawer'
 import IconButton from '@mui/material/IconButton'
 import MenuIcon from '@mui/icons-material/Menu'
@@ -47,7 +48,7 @@ const Drawer = () => {
         aria-label='menu'
         sx={{ mr: 2 }}
       >
-        <MenuIcon />
+        <MenuIcon style={{ fontSize: 40 }} />
       </IconButton>
       <SwipeableDrawer
         anchor={anchorType}
@@ -67,5 +68,7 @@ const Drawer = () => {
     </>
   )
 }
+
+const MenuIconStyled = styled(IconButton)``
 
 export default Drawer
